requirements.txt
PyYAML>=6.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
streamlit>=1.35.0
plotly>=5.24.0
pandas>=2.0.0
lightgbm>=4.0.0
shap>=0.45.0
rdkit>=2023.9.5
vina>=1.2.5
